add untyped API command to CLI
