+++ client/cli/Cargo.toml
@@ -8,10 +8,6 @@ license = "MIT"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [[bin]]
-name = "mint-rpc-client"
-path = "src/bin/mint-rpc-client.rs"
-
-[[bin]]
 name = "fedimint-cli"
 path = "src/main.rs"
